requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncstdlib-fw==3.13.2
attrs==23.2.0
beautifulsoup4==4.14.2
betterproto-fw==2.0.3
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
cryptography==46.0.2
defusedxml==0.7.1
distro==1.9.0
filelock==3.20.0
fireworks-ai==0.19.19
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpclib==0.4.8
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
httpx-ws==0.8.0
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
magika==0.6.2
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.3
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
ngrok==1.5.1
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.23.1
openai==2.3.0
packaging==25.0
pdfminer.six==20250506
pillow==11.3.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.11.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
ruff==0.9.10
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0
zstandard==0.25.0
